init `Card` works component
